chore: Move cross-arch SAE tests to correct file
